Keep track of what js_library code was generated for each proto_library.
Introduce a "TsProtoInfo" provider that keeps track of the ProtoInfo and JS
metadata for the ts_proto_library rule. By extracting the primary JS file
generated by a ts_proto_library rule, we can build a map of (proto import path
-> generated JS file) in Bazel.

The above mapping will next be used to rewrite the imports of the generated code
so that imports work properly. This is currently done in unpublished code using
a protoc plugin that dispatches to multiple other protoc plugins and then
rewrites their output... We can either do it that way or write a simple command
line tool that then gets called by the ts_proto_library rule.